segregate dependencies
